Pages Loaded and Choices Are Set
This is the completed version. If you go to MainActivity.java and
uncomment the code at the bottom, it will have the main page start
blank for the user to type in their name or a different name. I
commented it so that way the user name can stay up when reloaded.